% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{docsToNodes}
\alias{docsToNodes}
\title{Parse XML Documents into XML Nodes}
\usage{
docsToNodes(docs, xpath)
}
\arguments{
\item{docs}{XML documents}

\item{xpath}{xpath expression}
}
\description{
Essentially a recursive call to [XML::getNodeSet()].
}
