% Generated by roxygen2 (4.0.1): do not edit by hand
\name{docsToNodes}
\alias{docsToNodes}
\title{Parse XML Documents into XML Nodes}
\usage{
docsToNodes(docs, xpath)
}
\arguments{
\item{docs}{XML documents}

\item{xpath}{xpath expression}
}
\description{
Essentially a recursive call to \link{getNodeSet}.
}

