% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nodesToList}
\alias{nodesToList}
\title{Coerce XML Nodes into a list with both attributes and values}
\usage{
nodesToList(nodes)
}
\arguments{
\item{nodes}{A collection of XML nodes. Should be the output from \link{docsToNodes}.}
}
\value{
A nested list with a structure that resembles the XML structure
}
\description{
Essentially a recursive call to [XML::xmlToList()].
}
