% Generated by roxygen2 (4.0.1): do not edit by hand
\name{urlsToDocs}
\alias{urlsToDocs}
\title{Parse XML Files into XML Documents}
\usage{
urlsToDocs(urls, local = FALSE, quiet = FALSE, ...)
}
\arguments{
\item{urls}{character vector. Either urls that point to an XML file online or a local XML file name.}

\item{local}{logical. Should urls be treated as paths to local files?}

\item{quiet}{logical. Print file name currently being parsed?}

\item{...}{arguments passed along to \link{httr::GET}}
}
\description{
Essentially a recursive call to \link{xmlParse}.
}

