% Generated by roxygen2 (4.0.1): do not edit by hand
\name{urlsToDocs}
\alias{urlsToDocs}
\title{Parse XML Files into XML Documents}
\usage{
urlsToDocs(urls, quiet = FALSE, ...)
}
\arguments{
\item{urls}{character vector or list of urls that point to an XML file (or anything readable by \link{xmlParse}).}

\item{quiet}{logical. Print file name currently being parsed?}

\item{...}{arguments passed along to \link{httr::GET}}
}
\description{
Essentially a recursive call to \link{xmlParse}.
}

